+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - New NBT tag on transportalizers that allows them to be locked, permanently preventing the viewing or editing of its id or destination
 - Transportalizer NBT data will now be saved when mined, even without silk touch, and can be viewed on the item's tooltip
 - Added Carved Bush.
-- Scourging Heat computer Theme
+- New computer Themes: Astral Charts, Lifdoff, SBURB 10, Scourging Heat, Skaianet Black, Skaianet Green, Skaianet White, ~ATH
 
 ### Changed
 
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Computer Themes can now be added and rendered via resource pack
 - Default computer Theme for crockertop is now Crocker
 - Computer Theme is now chosen from a list of available Themes in separate GUI
+- Changes to computer Theme texture Crocker
 
 ### Fixed